Use of GeoTIFF files for georeferencing
The generation process for the heightmap and texture files was improved.  The new files were added to this commit.  The source files were too large to add to the commit.

A terrain.md file was added to describe the workflow.  It should be updated with the exact files used with URLs to locations to get them from the USGS.