Scaling correction and a name change
The "main" location was changed to the name of "blacksburg" as this is apparently the name the location goes by at VT.

There was a scaling issue with the maps.  The data is "1 m" data, but not precisely 1 m^2.  The 1028 square pixel map represent 2138 square meters of the physical world.  That is been changed (for both worlds).

Documentation was updated.